% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check.df}
\alias{check.df}
\title{Check data frame}
\usage{
check.df(x, column.names, column.conv = rep("as.character",
  length(column.names)), min.rows = 1)
}
\arguments{
\item{x}{The data.frame to be checked.}

\item{column.names}{character containing required column names}

\item{column.conv}{character containing conversion functions that should return TRUE for check to pass}

\item{min.rows}{integer specifying minimum number of rows of x}
}
\value{
The function returns a logical, whose value indicates pass (TRUE) or fail (FALSE).
}
\description{
check.df examines if a given data.frame has required columns corresponding to required classes.
}
\examples{
check.df(test.table, column.names = c("caseid", "age", "nh", "em", "aj"), column.conv = c("as.character", "as.integer", "as.logical", "as.logical", "as.logical"))
}
\author{
Johannes Elias
}
