% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsa_pdc.R
\name{mrsa.pdc}
\alias{mrsa.pdc}
\title{Prevalence dependent calibration of a model predicting probability of MRSA carriage and diagnostic test PPV}
\usage{
mrsa.pdc(tb, prev, spec, sens)
}
\arguments{
\item{tb}{A data.frame with the five columns 'caseid', 'age', 'nh', 'em', and 'aj'. Columns have to be of classes 'character', 'integer', 'logical', 'logical', and 'logical', respectively.}

\item{prev}{prevalence of MRSA, numeric ranging from 0 to 1}

\item{spec}{specificity of diagnostic test, numeric ranging from 0 to 1}

\item{sens}{sensitivity of diagnostic test, numeric ranging from 0 to 1}
}
\value{
The function returns data.frame tb with columns 'p' and 'PPV', corresponding to probability of MRSA carriage and estimated PPV of the diagnostic test, respectively, attached.
}
\description{
mrsa.pdc exposes a prevalence adjusted model which predicts probability of MRSA carriage positive predictive value (PPV) of a diagnostic test with known specificity and sensitivity.
}
\note{
The underlying model was used in an article by Elias et al (2013), available at https://doi.org/10.1186/1471-2334-13-111
}
\examples{
mrsa.pdc(test.table, .02, .98, .99)
}
\author{
Johannes Elias
}
